rename imports in service/controller/azure/v7patch1: github.com/giantswarm/cluster-operator/service/controller/azure/v7 > github.com/giantswarm/cluster-operator/service/controller/azure/v7patch1
