+++ service/controller/azure/v7patch1/resource/chartconfig/desired.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/giantswarm/cluster-operator/pkg/v7/chartconfig"
 	"github.com/giantswarm/cluster-operator/pkg/v7patch1/key"
-	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v7/key"
+	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v7patch1/key"
 )
 
 func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
